Move tests to root
