+++ jest.config.ts
@@ -1,0 +1,27 @@
+import type { Config } from '@jest/types';
+import { pathsToModuleNameMapper } from 'ts-jest/utils';
+import { compilerOptions } from './tsconfig.json';
+
+const config: Config.InitialOptions = {
+  roots: [
+    './website/examples',
+    './website/src',
+    './packages/react-day-picker/src'
+  ],
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest'
+  },
+  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, {
+    prefix: '<rootDir>'
+  }),
+  testEnvironment: 'jsdom',
+  coverageReporters: ['lcov', 'text', 'clover'],
+  setupFilesAfterEnv: ['./test/setup.ts'],
+  globals: {
+    'ts-jest': {
+      tsconfig: 'tsconfig.json'
+    }
+  }
+};
+
+export default config;
