--- packages/react-day-picker/jest.config.ts
@@ -1,13 +1,0 @@
-import type { Config } from '@jest/types';
-
-const config: Config.InitialOptions = {
-  roots: ['./src'],
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest'
-  },
-  testEnvironment: 'jsdom',
-  coverageReporters: ['lcov', 'text', 'clover'],
-  setupFilesAfterEnv: ['./src/test/setup.ts']
-};
-
-export default config;
