+++ packages/react-day-picker/package.json
@@ -21,7 +21,6 @@
     "build": "yarn prebuild && yarn rollup -c",
     "css-types": "yarn tcm -p style.css",
     "lint": "eslint --ext .ts,.tsx src",
-    "test": "jest",
     "typecheck": "tsc --project ./tsconfig.json --noEmit",
     "typecheck:watch": "tsc --project ./tsconfig.json --noEmit --skipLibCheck --watch"
   },
@@ -32,10 +31,6 @@
     "style.css.d.ts"
   ],
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.15.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@types/jest": "^27.0.3",
     "@types/node": "^14.17.34",
     "@types/react": "^17.0.37",
     "@typescript-eslint/eslint-plugin": "^5.5.0",
@@ -47,16 +42,12 @@
     "eslint-plugin-jest": "^25.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "jest": "^27.4.3",
-    "jest-date": "^1.1.4",
     "prettier": "^2.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "rimraf": "^3.0.2",
     "rollup": "^2.60.2",
     "rollup-plugin-typescript2": "^0.31.1",
-    "timekeeper": "^2.2.0",
-    "ts-jest": "^27.0.7",
     "ts-node": "^10.4.0",
     "tslib": "^2.3.1",
     "typed-css-modules": "^0.7.0",
