+++ packages/react-day-picker/tsconfig.json
@@ -97,5 +97,5 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "exclude": ["jest.config.ts", "build"]
+  "exclude": ["jest.config.ts", "build", "**/*.test.*"]
 }
