+++ website/package.json
@@ -9,7 +9,6 @@
     "deploy": "docusaurus deploy",
     "clear": "docusaurus clear",
     "lint": "eslint --ext .ts,.tsx src examples",
-    "test": "jest",
     "typecheck": "tsc --project ./tsconfig.json --noEmit --skipLibCheck",
     "typecheck:watch": "tsc --project ./tsconfig.json --noEmit --skipLibCheck --watch"
   },
@@ -24,12 +23,7 @@
     "@docusaurus/module-type-aliases": "^2.0.0-beta.9",
     "@docusaurus/preset-classic": "2.0.0-beta.9",
     "@docusaurus/theme-classic": "^2.0.0-beta.9",
-    "@testing-library/jest-dom": "^5.15.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@testing-library/user-event": "^13.5.0",
     "@tsconfig/docusaurus": "^1.0.4",
-    "@types/jest": "^27.0.3",
     "@types/react-helmet": "^6.1.4",
     "@types/react-router-dom": "^5.3.2",
     "@typescript-eslint/eslint-plugin": "^5.5.0",
@@ -42,11 +36,8 @@
     "eslint-plugin-jest": "^25.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "jest": "^27.4.3",
     "prettier": "^2.5.0",
     "raw-loader": "^4.0.2",
-    "timekeeper": "^2.2.0",
-    "ts-jest": "^27.0.7",
     "typedoc": "^0.22.10",
     "typedoc-plugin-markdown": "^3.11.7",
     "typescript": "~4.4.4",
